Shares Repository Implementation and the  data type of share_count row in the Share_Transfer table was changed from int to a decimal.

+++ SACCO System/Models/Shareholder.cs
@@ -9,7 +9,7 @@ public partial class Shareholder
 
     public int? MemberId { get; set; }
 
-    public int? ShareCount { get; set; }
+    public decimal? ShareCount { get; set; }
 
     public virtual Member? Member { get; set; }
 }
